Add a content page and allow navigation to it
Prompt: 1 Ok next we are going to make a "Content" page to add more media for users to review. We will also need a way to navigate to the new page. The content page will look similar, with cards.

Prompt: 2 Error in NavBar.tsx: You're importing a component that needs . This React hook only works in a client component. To fix, mark the file (or its parent) with the "use client" directive.